[orc8r][lib] Option to disable initflag flags parsing logic to accomodate some IDE testing/debugging environments
